# perf(dataset): improve performance on get list