--- a/tests/datasets/api_tests.py
+++ b/tests/datasets/api_tests.py
@@ -86,9 +86,12 @@ class DatasetApiTests(SupersetTestCase):
         response = json.loads(rv.data.decode("utf-8"))
         self.assertEqual(response["count"], 1)
         expected_columns = [
+            "database_id",
             "changed_by",
             "changed_by_name",
             "changed_by_url",
+            "changed_by_fk",
+            "default_endpoint",
             "changed_on",
             "database_name",
             "explore_url",