### CATEGORY

Choose one

- [ ] Bug Fix
- [ ] Enhancement (new features, refinement)
- [ ] Refactor
- [ ] Add tests
- [ ] Build / Development Environment
- [ ] Documentation

### SUMMARY
Like stated on #9619, when using properties has virtual columns, we should to make sure that the method references columns already referenced on list_columns or SQLAlchemy will issue a new query

Before:
[stats_logger] (timing) DatasetRestApi.get_list.time 110ms - 150ms

After:
[stats_logger] (timing) DatasetRestApi.get_list.time 25ms - 35ms

### ADDITIONAL INFORMATION
- [ ] Has associated issue:
- [ ] Changes UI
- [ ] Requires DB Migration.
- [ ] Confirm DB Migration upgrade and downgrade tested.
- [ ] Introduces new feature or API
- [ ] Removes existing feature or API

### REVIEWERS